feat: add tools exposed by enterprise so that specmatic mcp supports enterprise features
