+++ .gitignore
@@ -24,3 +24,4 @@ gen-resources/
 .kotlin/
 settings.local.json
 /temp
+.gemini
