+++ application/src/main/kotlin/application/mcp/server/McpToolProvider.kt
@@ -1,0 +1,8 @@
+package application.mcp.server
+
+import io.modelcontextprotocol.server.McpServerFeatures
+
+interface McpToolProvider {
+    val identifier: String
+    fun tools(): List<McpServerFeatures.SyncToolSpecification>
+}
