+++ application/src/main/resources/META-INF/services/application.mcp.server.McpToolProvider
@@ -1,0 +1,1 @@
+application.mcp.server.SpecmaticMcpToolProvider
